refactor: extract shared mock explore data into reusable module
