+++ packages/backend/src/ee/services/ai/utils/validateAxisFields.test.ts
@@ -1,129 +1,10 @@
 import {
-    DimensionType,
-    FieldType,
-    MetricType,
-    SupportedDbtAdapter,
-    type CustomMetricBase,
-    type Explore,
     type TableCalcsSchema,
     type ToolRunQueryArgsTransformed,
 } from '@lightdash/common';
 
+import { mockOrdersExplore } from './validationExplore.mock';
 import { validateAxisFields } from './validators';
-
-const mockExplore: Explore = {
-    targetDatabase: SupportedDbtAdapter.POSTGRES,
-    name: 'test_explore',
-    label: 'Test Explore',
-    tags: [],
-    spotlight: {
-        visibility: 'show',
-        categories: [],
-    },
-    baseTable: 'orders',
-    joinedTables: [],
-    tables: {
-        orders: {
-            name: 'orders',
-            label: 'Orders',
-            database: 'test_db',
-            schema: 'public',
-            sqlTable: 'orders',
-            sqlWhere: undefined,
-            uncompiledSqlWhere: undefined,
-            dimensions: {
-                order_date: {
-                    fieldType: FieldType.DIMENSION,
-                    type: DimensionType.DATE,
-                    name: 'order_date',
-                    label: 'Order Date',
-                    table: 'orders',
-                    tableLabel: 'Orders',
-                    sql: '${TABLE}.order_date',
-                    hidden: false,
-                    source: undefined,
-                    compiledSql: 'orders.order_date',
-                    tablesReferences: ['orders'],
-                },
-                customer_name: {
-                    fieldType: FieldType.DIMENSION,
-                    type: DimensionType.STRING,
-                    name: 'customer_name',
-                    label: 'Customer Name',
-                    table: 'orders',
-                    tableLabel: 'Orders',
-                    sql: '${TABLE}.customer_name',
-                    hidden: false,
-                    source: undefined,
-                    compiledSql: 'orders.customer_name',
-                    tablesReferences: ['orders'],
-                },
-                product_category: {
-                    fieldType: FieldType.DIMENSION,
-                    type: DimensionType.STRING,
-                    name: 'product_category',
-                    label: 'Product Category',
-                    table: 'orders',
-                    tableLabel: 'Orders',
-                    sql: '${TABLE}.product_category',
-                    hidden: false,
-                    source: undefined,
-                    compiledSql: 'orders.product_category',
-                    tablesReferences: ['orders'],
-                },
-            },
-            metrics: {
-                total_revenue: {
-                    fieldType: FieldType.METRIC,
-                    type: MetricType.SUM,
-                    name: 'total_revenue',
-                    label: 'Total Revenue',
-                    table: 'orders',
-                    tableLabel: 'Orders',
-                    sql: 'SUM(${TABLE}.amount)',
-                    hidden: false,
-                    source: undefined,
-                    compiledSql: 'SUM(orders.amount)',
-                    tablesReferences: ['orders'],
-                },
-                order_count: {
-                    fieldType: FieldType.METRIC,
-                    type: MetricType.COUNT,
-                    name: 'order_count',
-                    label: 'Order Count',
-                    table: 'orders',
-                    tableLabel: 'Orders',
-                    sql: 'COUNT(*)',
-                    hidden: false,
-                    source: undefined,
-                    compiledSql: 'COUNT(*)',
-                    tablesReferences: ['orders'],
-                },
-                avg_order_value: {
-                    fieldType: FieldType.METRIC,
-                    type: MetricType.AVERAGE,
-                    name: 'avg_order_value',
-                    label: 'Average Order Value',
-                    table: 'orders',
-                    tableLabel: 'Orders',
-                    sql: 'AVG(${TABLE}.amount)',
-                    hidden: false,
-                    source: undefined,
-                    compiledSql: 'AVG(orders.amount)',
-                    tablesReferences: ['orders'],
-                },
-            },
-            lineageGraph: {},
-            source: undefined,
-            groupLabel: undefined,
-        },
-    },
-    groupLabel: undefined,
-    warehouse: undefined,
-    sqlPath: undefined,
-    ymlPath: undefined,
-    databricksCompute: undefined,
-};
 
 describe('validateAxisFields', () => {
     describe('Edge Cases', () => {
