+++ packages/backend/src/ee/services/ai/utils/validateTableCalculations.test.ts
@@ -1,125 +1,21 @@
 import {
-    DimensionType,
-    FieldType,
     MetricType,
-    SupportedDbtAdapter,
     WindowFunctionType,
     type CustomMetricBase,
-    type Explore,
     type TableCalcsSchema,
 } from '@lightdash/common';
+import { mockOrdersExplore } from './validationExplore.mock';
 import { validateTableCalculations } from './validators';
-
-const mockExplore: Explore = {
-    targetDatabase: SupportedDbtAdapter.POSTGRES,
-    name: 'test_explore',
-    label: 'Test Explore',
-    tags: [],
-    spotlight: {
-        visibility: 'show',
-        categories: [],
-    },
-    baseTable: 'orders',
-    joinedTables: [],
-    tables: {
-        orders: {
-            name: 'orders',
-            label: 'Orders',
-            database: 'test_db',
-            schema: 'public',
-            sqlTable: 'orders',
-            sqlWhere: undefined,
-            uncompiledSqlWhere: undefined,
-            dimensions: {
-                order_id: {
-                    fieldType: FieldType.DIMENSION,
-                    type: DimensionType.NUMBER,
-                    name: 'order_id',
-                    label: 'Order ID',
-                    table: 'orders',
-                    tableLabel: 'Orders',
-                    sql: '${TABLE}.order_id',
-                    hidden: false,
-                    source: undefined,
-                    compiledSql: 'orders.order_id',
-                    tablesReferences: ['orders'],
-                },
-                customer_name: {
-                    fieldType: FieldType.DIMENSION,
-                    type: DimensionType.STRING,
-                    name: 'customer_name',
-                    label: 'Customer Name',
-                    table: 'orders',
-                    tableLabel: 'Orders',
-                    sql: '${TABLE}.customer_name',
-                    hidden: false,
-                    source: undefined,
-                    compiledSql: 'orders.customer_name',
-                    tablesReferences: ['orders'],
-                },
-            },
-            metrics: {
-                total_revenue: {
-                    fieldType: FieldType.METRIC,
-                    type: MetricType.SUM,
-                    name: 'total_revenue',
-                    label: 'Total Revenue',
-                    table: 'orders',
-                    tableLabel: 'Orders',
-                    sql: 'SUM(${TABLE}.amount)',
-                    hidden: false,
-                    source: undefined,
-                    compiledSql: 'SUM(orders.amount)',
-                    tablesReferences: ['orders'],
-                },
-                order_count: {
-                    fieldType: FieldType.METRIC,
-                    type: MetricType.COUNT,
-                    name: 'order_count',
-                    label: 'Order Count',
-                    table: 'orders',
-                    tableLabel: 'Orders',
-                    sql: 'COUNT(*)',
-                    hidden: false,
-                    source: undefined,
-                    compiledSql: 'COUNT(*)',
-                    tablesReferences: ['orders'],
-                },
-                max_date: {
-                    fieldType: FieldType.METRIC,
-                    type: MetricType.MAX,
-                    name: 'max_date',
-                    label: 'Max Date',
-                    table: 'orders',
-                    tableLabel: 'Orders',
-                    sql: 'MAX(${TABLE}.created_at)',
-                    hidden: false,
-                    source: undefined,
-                    compiledSql: 'MAX(orders.created_at)',
-                    tablesReferences: ['orders'],
-                },
-            },
-            lineageGraph: {},
-            source: undefined,
-            groupLabel: undefined,
-        },
-    },
-    groupLabel: undefined,
-    warehouse: undefined,
-    sqlPath: undefined,
-    ymlPath: undefined,
-    databricksCompute: undefined,
-};
 
 describe('validateTableCalculations', () => {
     describe('Edge Cases', () => {
         it('should not throw for null or empty tableCalcs array', () => {
             expect(() =>
-                validateTableCalculations(mockExplore, null, [], null),
+                validateTableCalculations(mockOrdersExplore, null, [], null),
             ).not.toThrow();
 
             expect(() =>
-                validateTableCalculations(mockExplore, [], [], null),
+                validateTableCalculations(mockOrdersExplore, [], [], null),
             ).not.toThrow();
         });
     });
@@ -139,7 +35,7 @@ describe('validateTableCalculations', () => {
 
             expect(() =>
                 validateTableCalculations(
-                    mockExplore,
+                    mockOrdersExplore,
                     tableCalcs,
                     selectedMetrics,
                     null,
@@ -161,7 +57,7 @@ describe('validateTableCalculations', () => {
 
             expect(() =>
                 validateTableCalculations(
-                    mockExplore,
+                    mockOrdersExplore,
                     tableCalcs,
                     selectedMetrics,
                     null,
@@ -196,7 +92,7 @@ describe('validateTableCalculations', () => {
 
             expect(() =>
                 validateTableCalculations(
-                    mockExplore,
+                    mockOrdersExplore,
                     tableCalcs,
                     selectedMetrics,
                     customMetrics,
@@ -226,7 +122,7 @@ describe('validateTableCalculations', () => {
 
             expect(() =>
                 validateTableCalculations(
-                    mockExplore,
+                    mockOrdersExplore,
                     tableCalcs,
                     selectedMetrics,
                     null,
@@ -255,7 +151,7 @@ describe('validateTableCalculations', () => {
 
             expect(() =>
                 validateTableCalculations(
-                    mockExplore,
+                    mockOrdersExplore,
                     tableCalcs,
                     selectedMetrics,
                     null,
@@ -281,7 +177,7 @@ describe('validateTableCalculations', () => {
 
             expect(() =>
                 validateTableCalculations(
-                    mockExplore,
+                    mockOrdersExplore,
                     tableCalcs,
                     selectedMetrics,
                     null,
@@ -314,7 +210,7 @@ describe('validateTableCalculations', () => {
 
             expect(() =>
                 validateTableCalculations(
-                    mockExplore,
+                    mockOrdersExplore,
                     tableCalcs,
                     selectedMetrics,
                     null,
@@ -346,7 +242,7 @@ describe('validateTableCalculations', () => {
 
             expect(() =>
                 validateTableCalculations(
-                    mockExplore,
+                    mockOrdersExplore,
                     tableCalcs,
                     selectedMetrics,
                     null,
@@ -374,7 +270,7 @@ describe('validateTableCalculations', () => {
 
             expect(() =>
                 validateTableCalculations(
-                    mockExplore,
+                    mockOrdersExplore,
                     tableCalcs,
                     selectedMetrics,
                     null,
@@ -398,7 +294,7 @@ describe('validateTableCalculations', () => {
 
             expect(() =>
                 validateTableCalculations(
-                    mockExplore,
+                    mockOrdersExplore,
                     tableCalcs,
                     selectedMetrics,
                     null,
@@ -420,7 +316,7 @@ describe('validateTableCalculations', () => {
 
             expect(() =>
                 validateTableCalculations(
-                    mockExplore,
+                    mockOrdersExplore,
                     tableCalcs,
                     selectedMetrics,
                     null,
@@ -443,7 +339,7 @@ describe('validateTableCalculations', () => {
 
             expect(() =>
                 validateTableCalculations(
-                    mockExplore,
+                    mockOrdersExplore,
                     tableCalcs,
                     selectedMetrics,
                     null,
@@ -466,7 +362,7 @@ describe('validateTableCalculations', () => {
 
             expect(() =>
                 validateTableCalculations(
-                    mockExplore,
+                    mockOrdersExplore,
                     tableCalcs,
                     selectedMetrics,
                     null,
@@ -491,7 +387,7 @@ describe('validateTableCalculations', () => {
 
             expect(() =>
                 validateTableCalculations(
-                    mockExplore,
+                    mockOrdersExplore,
                     tableCalcs,
                     selectedMetrics,
                     null,
